+++ src/Highcharts/Highchart.php
@@ -48,11 +48,13 @@ class Highchart extends AbstractChart
         $this->pane = ChartOption::instance('pane');
     }
 
+    #[\Override]
     protected function getChartClass(): string
     {
         return self::CHART_CLASS;
     }
 
+    #[\Override]
     protected function renderChartOptions(string &$chartJS): void
     {
         parent::renderChartOptions($chartJS);