+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added missing override attribute.
 - Updated chart option annotations.
 - Updated Psalm configuration. Reworked `HighchartsExtension` tests.
 - Updated tests.